+++ fitnessapp/templates/login.html
@@ -39,7 +39,7 @@
                 <a href="{{url_for('login')}}">Forgot Password?</a>
             </small>
             <small class="text-muted ml-2">
-                <a href="{{url_for('google_login')}}">Sign In With Google</a>
+                <a href="{{url_for('auth.google_login')}}">Sign In With Google</a>
             </small>
         </form>
     </div>
